move simple GeneMetadataModel logic
